refactor: split review and validate CLI commands into submodules
Apply the same treatment to the two next-largest CLI modules:

- req/src/cli/review.rs (647 lines) -> review.rs plus
  review/{display,accept}.rs: the CLI surface and shared filter logic
  stay in review.rs; output formatting and the accept workflow move to
  their own modules. The child/parent/kind filtering that was duplicated
  between display and accept modes is now a single apply_filters helper,
  and the identical JSON object construction in the json/ndjson formats
  is deduplicated into a link_json helper (with a display_path helper
  replacing six copies of the loaded-path-or-canonical-path fallback).

- req/src/cli/validate.rs (500 lines) -> validate.rs plus
  validate/{check,report,fix}.rs: the CLI surface and the result model
  stay in validate.rs; the health checks, output formatting, and
  automatic repair each get their own module.

No behaviour changes; all tests pass and clippy remains clean.
